Move all functions into new modular locations.
