+++ .pre-commit-config.yaml
@@ -1,9 +1,14 @@
 repos:
--   repo: https://github.com/psf/black
-    rev: 23.3.0
-    hooks:
-    -   id: black
--   repo: https://github.com/PyCQA/pylint
+  - repo: https://github.com/pycqa/pylint
     rev: v2.17.0
     hooks:
-    -   id: pylint
+      - id: pylint
+        args: [--rcfile=.pylintrc]
+  - repo: https://github.com/psf/black
+    rev: 23.3.0
+    hooks:
+      - id: black
+  - repo: https://github.com/pycqa/isort
+    rev: 5.12.0
+    hooks:
+      - id: isort
