+++ .pylintignore
@@ -1,0 +1,1 @@
+setup.py
