+++ docs/api/index.md
@@ -37,10 +37,11 @@ Tools for creating interactive visualizations.
 ## Common Workflows
 
 ### Basic Data Processing
+
 ```python
 from preprocessing.loading import XDrip
 from preprocessing.cleaning import clean_glucose, clean_classify_insulin, clean_classify_carbs
-from preprocessing.alignment import align_diabetes_data
+from core.alignment import align_diabetes_data
 
 # Load data
 data = XDrip('path_to_backup.sqlite')
