+++ setup.py
@@ -1,8 +1,21 @@
-from setuptools import setup, find_packages
+"""Basic setup file"""
+from setuptools import find_packages, setup
 
 setup(
-    name="cgm-data-processor",
+    name="cgm_data_processor",
     version="0.1",
-    packages=find_packages(where="src"),
-    package_dir={"": "src"},
+    packages=find_packages(),
+    install_requires=[
+        "pandas",
+        "numpy",
+        "sqlalchemy",
+        "plotly",
+    ],
+    extras_require={
+        "dev": [
+            "pylint",
+            "black",
+            "isort",
+        ]
+    },
 )
