+++ src/core/alignment.py
@@ -1,18 +1,13 @@
+"""Aligns diabetes data (blood glucose, carbohydrates, insulin) to regular 5-minute intervals
+    already established in the bg_df dataframe.
+    """
 import pandas as pd
-import numpy as np
-from typing import Dict, Any
-from datetime import datetime
 
 
 def align_diabetes_data(
-        bg_df: pd.DataFrame,
-        carb_df: pd.DataFrame,
-        insulin_df: pd.DataFrame
+    bg_df: pd.DataFrame, carb_df: pd.DataFrame, insulin_df: pd.DataFrame
 ) -> pd.DataFrame:
     """
-    Aligns diabetes data (blood glucose, carbohydrates, insulin) to regular 5-minute intervals
-    already established in the bg_df dataframe.
-
     This function takes three dataframes with timestamped diabetes data and produces a single
     dataframe with regular 5-minute intervals. Carbohydrates and insulin entries are summed. This
     dataset is suitable for ML, and other applications where a unified timestamp is required.
@@ -63,32 +58,37 @@ def align_diabetes_data(
     insulin_df = insulin_df.copy()
 
     # Round index timestamps to nearest 5 minute intervals
-    carb_df.index = carb_df.index.round('5min')
-    insulin_df.index = insulin_df.index.round('5min')
+    carb_df.index = carb_df.index.round("5min")
+    insulin_df.index = insulin_df.index.round("5min")
 
     # Resample and reindex insulin - sum within windows
-    insulin_resampled = insulin_df.resample('5min').agg({
-        'bolus': 'sum',
-        'basal': 'sum',
-        # Use a lambda to ensure empty groups are explicitly False - Empty groups equal True by
-        # default
-        'labeled_insulin': lambda x: x.all() if len(x) > 0 else False
-    }).reindex(bg_df.index)
+    insulin_resampled = (
+        insulin_df.resample("5min")
+        .agg(
+            {
+                "bolus": "sum",
+                "basal": "sum",
+                # Use a lambda to ensure empty groups are explicitly False - Empty groups equal True by
+                # default
+                "labeled_insulin": lambda x: x.all() if len(x) > 0 else False,
+            }
+        )
+        .reindex(bg_df.index)
+    )
 
     # Fill missing labeled_insulin values with False, explicitly astype boolean to avoid warnings
-    insulin_resampled['labeled_insulin'] = insulin_resampled['labeled_insulin'].astype('boolean').fillna(False)
+    insulin_resampled["labeled_insulin"] = (
+        insulin_resampled["labeled_insulin"].astype("boolean").fillna(False)
+    )
 
     # Resample and reindex carbs - sum within windows
-    carb_resampled = carb_df.resample('5min').agg({
-        'carbs': 'sum'
-    }).reindex(bg_df.index)
+    carb_resampled = carb_df.resample("5min").agg({"carbs": "sum"}).reindex(bg_df.index)
 
     # Combine all data
     aligned_df = pd.concat([bg_df, carb_resampled, insulin_resampled], axis=1)
 
     # Fill missing treatment values with 0
-    aligned_df['carbs'] = aligned_df['carbs'].fillna(0)
-    aligned_df[['basal', 'bolus']] = aligned_df[['basal', 'bolus']].fillna(0)
-
+    aligned_df["carbs"] = aligned_df["carbs"].fillna(0)
+    aligned_df[["basal", "bolus"]] = aligned_df[["basal", "bolus"]].fillna(0)
 
     return aligned_df
