+++ src/visualisation/dashboards/gaps.py
@@ -1,13 +1,18 @@
-import pandas as pd
+"""Creates an interactive dashboard visualizing glucose monitoring data gaps."""
+from typing import Dict, Optional
+
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
-from typing import Dict
 
 
-def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.png", display: bool = True,
-                         width: int = 1000, height: int = 1900) -> Optional[go.Figure]:
-    """Creates an interactive dashboard visualizing glucose monitoring data gaps.
-
+def create_gap_dashboard(
+    gaps_data: Dict,
+    save_path: str = "img/gaps_dashboard.png",
+    display: bool = True,
+    width: int = 1000,
+    height: int = 1900,
+) -> Optional[go.Figure]:
+    """
     Generates a comprehensive dashboard using Plotly with multiple visualizations
     showing patterns and statistics of missing glucose readings.
 
@@ -58,35 +63,35 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         ...     height=2000
         ... )
         >>> fig.show()  # Display interactive dashboard
-        Enhanced Gaps Dashboard saved to glucose_gaps.png
-"""
+        Enhanced Gaps Dashboard saved to glucose_gaps.png"""
     # Extract the gaps data
-    largest_gaps = gaps_data.get('largest_gaps', None)
-    gaps_df = gaps_data.get('gaps_df', None)
+    largest_gaps = gaps_data.get("largest_gaps", None)
+    gaps_df = gaps_data.get("gaps_df", None)
 
     # New statistics
-    initial_missing_percentage = gaps_data.get('initial_missing_percentage', 100)
-    initial_missing_count = gaps_data.get('initial_missing_count', 0)
-    remaining_missing_percentage = gaps_data.get('remaining_missing_percentage', 100)
-    remaining_missing_count = gaps_data.get('remaining_missing_count', 0)
-    total_readings = gaps_data.get('total_readings', 0)
-    total_gap_minutes = gaps_data.get('total_gap_minutes', 0)
-    average_gap_minutes = gaps_data.get('average_gap_minutes', 0)
-    median_gap_minutes = gaps_data.get('median_gap_minutes', 0)
+    initial_missing_percentage = gaps_data.get("initial_missing_percentage", 100)
+    initial_missing_count = gaps_data.get("initial_missing_count", 0)
+    remaining_missing_percentage = gaps_data.get("remaining_missing_percentage", 100)
+    remaining_missing_count = gaps_data.get("remaining_missing_count", 0)
+    total_readings = gaps_data.get("total_readings", 0)
+    total_gap_minutes = gaps_data.get("total_gap_minutes", 0)
+    average_gap_minutes = gaps_data.get("average_gap_minutes", 0)
+    median_gap_minutes = gaps_data.get("median_gap_minutes", 0)
 
     # Calculate additional descriptive statistics
     if gaps_df is not None:
-        q1_gap = gaps_df['length_minutes'].quantile(0.25)
-        q3_gap = gaps_df['length_minutes'].quantile(0.75)
-        std_dev_gap = gaps_df['length_minutes'].std()
-        min_gap = gaps_df['length_minutes'].min()
-        max_gap = gaps_df['length_minutes'].max()
+        q1_gap = gaps_df["length_minutes"].quantile(0.25)
+        q3_gap = gaps_df["length_minutes"].quantile(0.75)
+        std_dev_gap = gaps_df["length_minutes"].std()
+        min_gap = gaps_df["length_minutes"].min()
+        max_gap = gaps_df["length_minutes"].max()
     else:
         q1_gap = q3_gap = std_dev_gap = min_gap = max_gap = None
 
     # Create subplots with new layout
     fig = make_subplots(
-        rows=4, cols=2,
+        rows=4,
+        cols=2,
         specs=[
             [{"type": "indicator"}, {"type": "indicator"}],
             [{"type": "indicator"}, {"type": "indicator"}],
@@ -94,11 +99,14 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
             [{"type": "table", "colspan": 2}, None],
         ],
         subplot_titles=(
-            "", "",
-            "", "",
-            "Top 10 Largest Gaps", "All Gaps Distribution",
-            "Comprehensive Statistics"
-        )
+            "",
+            "",
+            "",
+            "",
+            "Top 10 Largest Gaps",
+            "All Gaps Distribution",
+            "Comprehensive Statistics",
+        ),
     )
 
     # Indicator for total gaps before interpolation
@@ -106,11 +114,12 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         go.Indicator(
             mode="gauge+number",
             value=initial_missing_count,
-            title={'text': "Initial Gaps"},
-            number={'suffix': " Gaps"},
-            gauge={'axis': {'range': [None, initial_missing_count]}},
+            title={"text": "Initial Gaps"},
+            number={"suffix": " Gaps"},
+            gauge={"axis": {"range": [None, initial_missing_count]}},
         ),
-        row=1, col=1
+        row=1,
+        col=1,
     )
 
     # Indicator for initial missing percentage
@@ -118,11 +127,12 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         go.Indicator(
             mode="gauge+number",
             value=initial_missing_percentage,
-            title={'text': "Initial Missing Data"},
-            number={'suffix': "%"},
-            gauge={'axis': {'range': [0, 100]}},
+            title={"text": "Initial Missing Data"},
+            number={"suffix": "%"},
+            gauge={"axis": {"range": [0, 100]}},
         ),
-        row=1, col=2
+        row=1,
+        col=2,
     )
 
     # Indicator for total gaps before interpolation
@@ -130,11 +140,12 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         go.Indicator(
             mode="gauge+number",
             value=remaining_missing_count,
-            title={'text': "Remaining Gaps"},
-            number={'suffix': " Gaps"},
-            gauge={'axis': {'range': [None, (remaining_missing_count*2.2)]}},
+            title={"text": "Remaining Gaps"},
+            number={"suffix": " Gaps"},
+            gauge={"axis": {"range": [None, (remaining_missing_count * 2.2)]}},
         ),
-        row=2, col=1
+        row=2,
+        col=1,
     )
 
     # Indicator for initial missing percentage
@@ -142,38 +153,41 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         go.Indicator(
             mode="gauge+number",
             value=remaining_missing_percentage,
-            title={'text': "Remaining Missing Data"},
-            number={'suffix': "%"},
-            gauge={'axis': {'range': [0, 100]}},
+            title={"text": "Remaining Missing Data"},
+            number={"suffix": "%"},
+            gauge={"axis": {"range": [0, 100]}},
         ),
-        row=2, col=2
+        row=2,
+        col=2,
     )
 
     # Bar chart for largest gaps
     if largest_gaps is not None:
         fig.add_trace(
             go.Bar(
-                x=largest_gaps['start_time'],
-                y=largest_gaps['length_minutes'],
+                x=largest_gaps["start_time"],
+                y=largest_gaps["length_minutes"],
                 name="Largest Gaps",
-                marker_color='orange',
-                hovertemplate="Start: %{x}<br>Duration: %{y} minutes<extra></extra>"
+                marker_color="orange",
+                hovertemplate="Start: %{x}<br>Duration: %{y} minutes<extra></extra>",
             ),
-            row=3, col=1
+            row=3,
+            col=1,
         )
 
     # Scatter plot for all gaps
     if gaps_df is not None:
         fig.add_trace(
             go.Scatter(
-                x=gaps_df['start_time'],
-                y=gaps_df['length_minutes'],
-                mode='markers',
-                marker=dict(color='blue', opacity=0.6),
+                x=gaps_df["start_time"],
+                y=gaps_df["length_minutes"],
+                mode="markers",
+                marker=dict(color="blue", opacity=0.6),
                 name="All Gaps",
-                hovertemplate="Time: %{x}<br>Duration: %{y} minutes<extra></extra>"
+                hovertemplate="Time: %{x}<br>Duration: %{y} minutes<extra></extra>",
             ),
-            row=3, col=2
+            row=3,
+            col=2,
         )
 
     # Comprehensive statistics table
@@ -181,9 +195,9 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
         go.Table(
             header=dict(
                 values=["Metric", "Value"],
-                fill_color='paleturquoise',
-                align='left',
-                font=dict(size=12)
+                fill_color="paleturquoise",
+                align="left",
+                font=dict(size=12),
             ),
             cells=dict(
                 values=[
@@ -200,7 +214,7 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
                         "75th Percentile Gap",
                         "Standard Deviation",
                         "Minimum Gap",
-                        "Maximum Gap"
+                        "Maximum Gap",
                     ],
                     [
                         f"{total_readings:,}",
@@ -215,24 +229,25 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
                         f"{q3_gap:.1f} minutes",
                         f"{std_dev_gap:.1f} minutes",
                         f"{min_gap:.1f} minutes",
-                        f"{max_gap:.1f} minutes"
-                    ]
+                        f"{max_gap:.1f} minutes",
+                    ],
                 ],
-                align='center',
-                font=dict(size=10)
-            )
+                align="center",
+                font=dict(size=10),
+            ),
         ),
-        row=4, col=1
+        row=4,
+        col=1,
     )
 
     # Update layout with new dimensions and better use of space
     fig.update_layout(
         title={
-            'text': "Glucose Reading Gaps Analysis Dashboard",
-            'y': 0.98,
-            'x': 0.5,
-            'xanchor': 'center',
-            'yanchor': 'top'
+            "text": "Glucose Reading Gaps Analysis Dashboard",
+            "y": 0.98,
+            "x": 0.5,
+            "xanchor": "center",
+            "yanchor": "top",
         },
         showlegend=False,
         height=height,
@@ -244,8 +259,8 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
     )
 
     # Update font sizes for all subplot titles
-    for annotation in fig['layout']['annotations']:
-        annotation['font'] = dict(size=12)
+    for annotation in fig["layout"]["annotations"]:
+        annotation["font"] = dict(size=12)
 
     # Update table font sizes
     for trace in fig.data:
@@ -265,3 +280,5 @@ def create_gap_dashboard(gaps_data: Dict, save_path: str = "img/gaps_dashboard.p
     # Return figure for display if requested
     if display:
         return fig
+
+    return None
